==> analyses_for_figures/figures_3_4_Müllerian_Wolffian_regionalisation/3a_Müllerian_Duct_Differentiation_Mesenchyme.ipynb <==
## Müllerian duct differentiation - mesenchyme

import numpy as np
import pandas as pd
import scanpy as sc
import scrublet as scr
import seaborn as sns
import scipy.stats
import anndata
import os

import scipy as scipy
import scipy as sp
import pickle as pkl
import matplotlib.pyplot as plt
import re
from collections import defaultdict
from statsmodels.nonparametric.smoothers_lowess import lowess
from numpy import asarray as ar
from collections import Counter
import networkx as nx
import igraph
import glob

print(anndata.__version__)

import matplotlib as mpl
mpl.rcParams['pdf.fonttype'] = 42

females_post9pcw = sc.read('/nfs/team292/vl6/FetalReproductiveTract/post_10pcw_females.20240326.h5ad')
females_post9pcw.shape

sc.set_figure_params(scanpy=True, dpi=80, dpi_save=150, frameon=True, vector_friendly=True, 
                         fontsize=14, figsize=(7, 7), color_map=None, format='pdf', facecolor=None, transparent=False)

sc.pl.umap(females_post9pcw, color="celltype")

sc.pl.umap(females_post9pcw, color="celltype", groups = ['FallopianTube_Mesenchyme', 
                'Uterus_Mesenchyme', 'Cervix_Mesenchyme', 'MüllerianVagina_Mesenchyme'
                                                                         ])

females_post9pcw = females_post9pcw[[i in ['FallopianTube_Mesenchyme', 
                'Uterus_Mesenchyme', 'Cervix_Mesenchyme', 'MüllerianVagina_Mesenchyme']
                                    for i in females_post9pcw.obs['celltype']]]

sc.pl.umap(females_post9pcw, color="celltype")

import anndata

adata = anndata.AnnData(X = females_post9pcw.raw.X, obs = females_post9pcw.obs, var = females_post9pcw.raw.var)
adata

adata = adata[[i == 'G1' for i in adata.obs['phase']]]

adata.obs['stage_pcw'].value_counts(dropna = False)

# Random downsampling per cell type 
import random
from itertools import chain
def downsample(adata, labels, n): 
    
    myindex = adata.obs[labels].value_counts().index 
    myvalues = adata.obs[labels].value_counts().values
    clusters = pd.Series(myvalues, index = myindex)
    
    # Find clusters with > n cells 
    cl2downsample = clusters.index[ clusters.values > n ]

    # save all barcode ids from small clusters
    holder = []
    holder.append( adata.obs_names[[ i not in cl2downsample for i in adata.obs[labels] ]] ) 

    # randomly sample n cells in the cl2downsample
    for cl in cl2downsample:
        print(cl)
        cl_sample = adata[[ i == cl for i in adata.obs[labels]]].obs_names
        # n = int(round(len(cl_sample)/2, 0))
       
        cl_downsample = random.sample(set(cl_sample), n )
        holder.append(cl_downsample)
    
    # samples to include
    samples = list(chain(*holder))

    # Filter adata_count
    adata = adata[[ i in samples for i in adata.obs_names ]]
    return adata

## import reproductive tract utils functions
import sys
cwd = '/nfs/team292/vl6/Experiments/Utils/'
sys.path.append(cwd)

import reptract_utils
import reptract_genes

adata.shape

adata.raw = adata.copy()

sc.pp.filter_genes(adata, min_counts=10)

adata = reptract_utils.normalize_log_transform(adata)

adata = reptract_utils.hvgs_pca_umap(adata)

sc.pl.umap(adata, color=['stage_pcw', 'sample', 
                         'n_genes', 'donor',  'doublet_scores',
                        'celltype', 'phase'], ncols = 2, wspace = 0.3, 
          color_map = 'OrRd')

sc.pl.umap(adata, color = [ 'FOXL2','ITGA4', 'LGR5', 'TMEM176B', 'HOXA10', 'PTGER3', 'ACTA2', 'DES', 'MYH11',
                          'HOXA13', 'ISL1', 'TMEM163', 'ZAP70', 'ITGBL1',
                           'MMP28', 'CTNNB1',  'WNT4', 'CDH7', 'ATF3', 'HMGA2', 
                          'KRT18', 'HOXA11', 'HOXC10', 'SRD5A2', 'EMX2', 'FOXF1', 'FOXF2'], wspace = 0.15, use_raw = False,
           color_map = 'OrRd', ncols = 2)

sc.tl.leiden(adata, resolution = 5)
sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data')

sc.tl.leiden(adata, resolution = 0.6, restrict_to = ('leiden', ['8', '20', '5']))
sc.pl.umap(adata, color = 'leiden_R')

sc.pl.umap(adata, color = 'leiden_R', groups = ['55', '28', '62', '42', '8-20-5,3', '8-20-5,5', 
                                               '8-20-5,4', '52'])

sc.tl.leiden(adata, resolution = 1, restrict_to = ('leiden_R', ['58', '23']), key_added = 'leiden_2')
sc.pl.umap(adata, color = 'leiden_2')

sc.pl.umap(adata, color = 'leiden_2', groups = ['55', '28', '62', '42', '8-20-5,3', '8-20-5,5', 
                                               '8-20-5,', '52', '56', '58-23,5', '58-23,1', '58-23,8'])

adata = adata[[i not in ['55', '28', '62', '42', '8-20-5,3', '8-20-5,5', 
                    '8-20-5,', '52', '56', '58-23,5', '58-23,1', '58-23,8'] for i in adata.obs['leiden_2']]]

sc.pl.umap(adata, color = [ 'FOXL2','ITGA4', 'LGR5', 'TMEM176B', 'HOXC10', 'HOXA10', 'TP63', 'PTGER3', 'ACTA2', 'DES', 'MYH11', 
                          'HOXA13', 'ISL1', 'TMEM163', 'MMP28', 'SRD5A2'], wspace = 0.15, use_raw = False,
           color_map = 'OrRd', ncols = 2)

adata.obs['stage_pcw'].value_counts()

# adata = downsample(adata, 'stage_pcw', 3000)

adata = anndata.AnnData(X = adata.raw.X, var = adata.raw.var, obs = adata.obs)


adata.raw = adata.copy()
sc.pp.filter_genes(adata, min_counts=10)
#adata = reptract_utils.per_gene_analysis(adata)
adata = reptract_utils.normalize_log_transform(adata)
adata = reptract_utils.hvgs_pca_umap(adata)

sc.pl.umap(adata, color=['stage_pcw', 'sample', 
                         'n_genes', 'donor',  'doublet_scores',
                        'celltype'], ncols = 2, wspace = 0.3, 
          color_map = 'OrRd')

sc.pl.umap(adata, color = [ 'FOXL2','ITGA4', 'LGR5', 'TMEM176B',  'HOXA10',  'PTGER3', 'CDH7', 
                          'MYH11', 'HOXA13', 'HOXC10', 'ACTA2', 'MYH11', 'CNTN1', 'EMC1', 'TMEM163', 
                          'MMP28', 'HOXC10', 'SRD5A2', 'EMX2', 'FOXF1', 'FOXF2'], wspace = 0.15, use_raw = False,
           color_map = 'OrRd', ncols = 2)

sc.tl.leiden(adata, resolution = 7)
sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data')

# sc.tl.leiden(adata, resolution = 0.2, restrict_to = ('leiden', ['30']))
# sc.pl.umap(adata, color = 'leiden_R')

sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data', groups = ['89', '58', '77', '90', '9', '64',
                                                                     '55', '78', '74'])

adata = adata[[i not in ['89', '58', '77', '90', '9', '64',
                                                                     '55', '78', '74'] for i in adata.obs['leiden']]]

sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data')

adata = anndata.AnnData(X = adata.raw.X, var = adata.raw.var, obs = adata.obs)
adata.raw = adata.copy()
sc.pp.filter_genes(adata, min_counts=10)
#adata = reptract_utils.per_gene_analysis(adata)
adata = reptract_utils.normalize_log_transform(adata)
adata = reptract_utils.hvgs_pca_umap(adata)

sc.pl.umap(adata, color = [ 'FOXL2','ITGA4', 'LGR5', 'TMEM176B',  'HOXA10',  'PTGER3', 'CDH7', 
                          'MYH11', 'HOXA13', 'HOXC10', 'FOXF1', 'EMX2'], wspace = 0.15, use_raw = False,
           color_map = 'OrRd', ncols = 2)

sc.pl.umap(adata, color=['stage_pcw', 'sample', 
                         'n_genes', 'donor',  'doublet_scores',
                        'celltype'], ncols = 2, wspace = 0.3, 
          color_map = 'OrRd')

import rpy2.rinterface_lib.callbacks
import logging
# Ignore R warning messages
#Note: this can be commented out to get more verbose R output
rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR)
import anndata2ri
anndata2ri.activate()
%load_ext rpy2.ipython

# Extract PCA matrix and batch array
n_pcs = 15 # Principal components used to perform dimensionality reduction 
pca = adata.obsm['X_pca'][:, 0:(n_pcs)]
batch = adata.obs['donor'] # Batch ID, in this case stage

%%R -i pca -i batch -o hem

library(harmony)
library(magrittr)
set.seed(1000)
hem <- HarmonyMatrix(pca, batch, theta=0, lambda = 6, do_pca=FALSE, verbose = FALSE)  
hem = data.frame(hem)

# Add harmony values to the anndata object --> overwriting PCA 
adata.obsm['X_pca'] = hem.values 

sc.pp.neighbors(adata, n_pcs = n_pcs, random_state= 123)
sc.tl.umap(adata, min_dist=0.4, spread = 1.2)

sc.pl.umap(adata, color = [ 'FOXL2','ITGA4', 'LGR5', 'TMEM176B','HOXA10', 'ESR1', 'PGR', 
                          'HOXA13', 'MYH11', 'PTGER3', 'CDH7', 'HOXC10', 'EMX2', 'SRD5A2', 'HOXA13'], wspace = 0.15, use_raw = False,
           color_map = 'OrRd', ncols = 2)

sc.pl.umap(adata, color=['stage_pcw', 
                         'n_genes', 'donor',  'doublet_scores',
                        'celltype', 'phase'], ncols = 2, wspace = 0.3, 
          color_map = 'OrRd')

sc.tl.leiden(adata, resolution = 0.7)
sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data')

sc.tl.draw_graph(adata)

sc.pl.draw_graph(adata, color='leiden', legend_loc='on data')


sc.pl.draw_graph(adata, color='n_genes', color_map='OrRd')


sc.tl.paga(adata, groups='leiden')


sc.pl.paga(adata, color=['leiden', 'TMEM163', 'LGR5', 'ITGA4', 'HOXA13', 'SRD5A2'])


sc.tl.draw_graph(adata, init_pos='paga')


sc.pl.draw_graph(adata, color='leiden', legend_loc='on data')


adata

import anndata
bdata = anndata.AnnData(X = adata.raw.X, var = adata.raw.var, obs = adata.obs)

import rpy2.rinterface_lib.callbacks
import logging
# Ignore R warning messages
#Note: this can be commented out to get more verbose R output
rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR)
import anndata2ri
anndata2ri.activate()
%load_ext rpy2.ipython

for col in bdata.obs.columns: 
    if col != 'leiden':
        del bdata.obs[col]

%%R -i bdata
bdata

%%R -o mrks

library(SoupX)
counts <- assay(bdata, "X")
colnames(counts) <- colnames(bdata)
rownames(counts) <- rownames(bdata)
mrks = quickMarkers(counts, colData(bdata)$leiden, N = 50) # where clusters is the cell type assignment and 100 means I want the top 100 genes per cluster that pass the hypergeometric test

sc.pl.umap(adata, color = ['leiden', 'stage_pcw'], legend_loc = 'on data')

cluster = mrks[mrks['cluster'] == '0']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '1']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '2']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '3']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '4']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '5']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '6']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '7']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '8']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '9']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

cluster = mrks[mrks['cluster'] == '10']['gene'].to_list()[0:9]
sc.pl.umap(adata, color = cluster, color_map = 'OrRd', ncols = 3, use_raw = False)

sc.pl.umap(adata, color = 'leiden', legend_loc = 'on data')

anno1 = {'0' : 'FallopianTube_Mesenchyme_late', '1' : 'Uterus_Mesenchyme_late', '2' : 'Cervix_Mesenchyme_late',
         '3' : 'Uterus_Mesenchyme_late', '4' : 'MüllerianVagina_Mesenchyme','5' : 'FallopianTube_Mesenchyme_early',
         '6' : 'Uterus_Mesenchyme_early', '7' : 'FallopianTube_Mesenchyme_late', '8' : 'FallopianTube_Mesenchyme_early',
        '9' : 'Cervix_Mesenchyme_early', '10' : 'Uterus_Mesenchyme_late',}

anno2 = {'0' : 'FallopianTube_Mesenchyme', '1' : 'Uterus_Mesenchyme', '2' : 'Cervix_Mesenchyme',
         '3' : 'Uterus_Mesenchyme', '4' : 'MüllerianVagina_Mesenchyme','5' : 'FallopianTube_Mesenchyme',
         '6' : 'Uterus_Mesenchyme', '7' : 'FallopianTube_Mesenchyme', '8' : 'FallopianTube_Mesenchyme',
        '9' : 'Cervix_Mesenchyme', '10' : 'Uterus_Mesenchyme'}


adata.obs['mese_mullerian_lowres'] = adata.obs.leiden.map(anno2)
adata.obs['mese_mullerian_highres'] = adata.obs.leiden.map(anno1)

sc.pl.umap(adata, color = ['stage_pcw', 'donor', 'leiden', 'mese_mullerian_lowres', 'mese_mullerian_highres'], 
          ncols = 2, save = '_mese_mullerian')

#del adata.obs['epi_mullerian_lowres']
#del adata.obs['epi_mullerian_highres']

adata.write('/nfs/team292/vl6/FetalReproductiveTract/mullerian_mese_late_post10pcw.h5ad')

adata = sc.read('/nfs/team292/vl6/FetalReproductiveTract/mullerian_mese_late_post10pcw.h5ad')

adata.shape

adata.obs['mese_mullerian_lowres'] = adata.obs['mese_mullerian_lowres'].astype('category')
adata.obs['mese_mullerian_lowres'] = adata.obs['mese_mullerian_lowres'].cat.reorder_categories(
[ 'Mesenchymal_FallopianTube', 'Mesenchymal_Uterus', 'Mesenchymal_Cervix', 
'Mesenchymal_UpperVagina'])

sc.pl.umap(adata, color = ['stage', 'mese_mullerian_lowres'], palette = ['blueviolet', 'navy', 
                                                                         'deepskyblue',
                                                                        'powderblue'],
          ncols = 2, save = '_mese_mullerian_lowres')

sc.pl.umap(adata, color = ['mese_mullerian_lowres'], palette = ['blueviolet', 'navy', 
                                                                         'deepskyblue',
                                                                        'powderblue'],
          ncols = 1, save = '_mese_mullerian_lowres_celltype')

sc.pl.umap(adata, color = ['stage'],
          ncols = 1, save = '_mese_mullerian_lowres_stage')

sc.pl.umap(adata, color = ['donor'],
          ncols = 1, save = '_mese_mullerian_lowres_donor')

adata.obs['mese_mullerian_highres'] = adata.obs['mese_mullerian_highres'].astype('category')
adata.obs['mese_mullerian_highres'] = adata.obs['mese_mullerian_highres'].cat.reorder_categories(
['Mesenchymal_FallopianTube_early', 'Mesenchymal_FallopianTube_late', 
 'Mesenchymal_Uterus_early', 'Mesenchymal_Uterus_late', 'Mesenchymal_Cervix', 'Mesenchymal_UpperVagina'])

sc.pl.umap(adata, color = ['stage', 'mese_mullerian_highres'], palette = [ 'goldenrod',
                                                                         'darkorange', 'deepskyblue', 'dodgerblue',
                                                                        'yellowgreen', 'forestgreen'],
          ncols = 2, save = '_mese_mullerian_highres')

sc.pl.umap(adata, color = ['stage', 'mese_mullerian_highres'], palette = [ 'goldenrod',
                                                                         'darkorange', 'deepskyblue', 'dodgerblue',
                                                                        'yellowgreen', 'forestgreen'],
          ncols = 2)

sc.pl.umap(adata, color = ['NRK', 'NRG1', 'GATA6', 'FOXL2', 'EMX2', 'ESR1', 'RORB', 'ESRRG', 'SFRP5', 
                          'PBX3', 'PRRX2', 'AHR', 'SALL3', 'ISL1', 'TCF21', 'NFATC2', 'PROX1', 
                          'BNC2', 'GATA6', 'RARB', 'ETV1', 'ETS1', 'ETV5', 'EVX1', 'EVX2', 'LBX2', 'SP9', 'ELK3',
                          'TFAP2C', 'SOX5'],
           use_raw = False, color_map = 'OrRd',
          ncols = 2)
