==> analyses_for_figures/figures_3_4_Müllerian_Wolffian_regionalisation/6a_MüllerianAxis_Mesenchymal_Spots_TradeSeq.ipynb <==
## tradeSeq for DE along Mullerian rostro-caudal axis in female reproductive tract

#if (!require("BiocManager", quietly = TRUE))
#    install.packages("BiocManager")

#BiocManager::install("tradeSeq")

#install.packages('igraph')

ip = as.data.frame(installed.packages()[,c(1,3:4)])
ip = ip[is.na(ip$Priority),1:2,drop=FALSE]
head(ip)

library(tradeSeq)
library(SingleCellExperiment)
library(slingshot)

library(SoupX)

sessionInfo()

adata <- readRDS("/lustre/scratch126/cellgen/team292/vl6/VISIUM/femalereproductiveaxis_mese_cytassist_downsampled.rds")
adata


counts <- assay(adata, "X")

set.seed(7)
pseudotime <- colData(adata)$FemaleReproductiveAxis
cellWeights <- rep(1, 501)

#icMat <- evaluateK(counts = counts, sds = SlingshotDataSet(adata), k = 3:10, 
#                   nGenes = 200, verbose = T)

#BiocManager::install("scran")

library(scran)

libsizes <- colSums(counts)
size.factors <- libsizes/mean(libsizes)
logcounts(adata) <- log2(t(t(counts)/size.factors) + 1)
assayNames(adata)

# Filter genes with non-zero biological components
getVariableGenes <- function(adata,fdr_thresh=NULL,var_thresh=0) {
  gene_var <- modelGeneVar(adata)
  genes <- getTopHVGs(gene_var, var.threshold =var_thresh ,fdr.threshold = fdr_thresh)
  return(genes)
}

genes <- getVariableGenes(adata)

length(genes)

dim(counts[genes,])

sce <- fitGAM(counts = counts[genes,], pseudotime = pseudotime, cellWeights = cellWeights, nknots = 10,
                verbose = TRUE, parallel = TRUE)

table(rowData(sce)$tradeSeq$converged)

adata <- adata[genes, ]

adata

gamList <- fitGAM(counts = counts[genes,], pseudotime = pseudotime, cellWeights = cellWeights, nknots = 10,
                verbose = TRUE, 
                  parallel = TRUE, sce = FALSE)

summary(gamList[["SRD5A2"]])


pvalLineage <- getSmootherPvalues(gamList)
statLineage <- getSmootherTestStats(gamList)

### ASSOCIATION TEST

A first exploration of the data analysis may consist of checking whether gene expression is associated with a particular lineage. The statistical test performed here, implemented in the associationTest function, is testing the null hypothesis that all smoother coefficients are equal to each other. This can be interpreted as testing whether the average gene expression is significantly changing along pseudotime.



assoRes <- associationTest(sce)
head(assoRes)

write.csv(assoRes, file = '/nfs/team292/vl6/FetalReproductiveTract/manuscript_v1/supp_tables/SupplementaryTable8_müllerian_mese_cytassist.csv')


### FILTER GENES AND CLUSTER THEM BY PATTERN

dim(assoRes)

assoRes = assoRes[assoRes$pvalue < 0.001,]
dim(assoRes)

assoRes = assoRes[assoRes$waldStat > 500,]
dim(assoRes)

assoRes['HOXA13',]

assoRes = assoRes[order(assoRes$waldStat, decreasing = TRUE),]

tail(assoRes)

# Get the index of the DataFrame
tops <- row.names(assoRes)[1:50]

tops

library(ggplot2)

for (gene in tops) {
    print(gene)
    print(plotSmoothers(sce, counts, gene = gene, alpha = 0) + ggtitle(gene) +
                          theme(title = element_text(size = 20, face = 'bold'),
                                axis.title = element_text(size = 20),
                                axis.text = element_text(size = 20),
                                legend.text = element_text(size = 20),
                                legend.title = element_text(size = 20, face = "bold")))
}

sce

model_vals_df <- predictSmooth(models = sce, gene=row.names(assoRes), nPoints=100) # output is 715 genes x 100 pseudospace values 

model_vals_mtx <- predictSmooth(models = sce, gene=row.names(assoRes), nPoints=100, tidy=FALSE)

dim(model_vals_mtx) # 715 genes x 100 pseudospace values 

model_vals_mtx[1:10, 1:20]

write.csv(model_vals_mtx, file = '/lustre/scratch126/cellgen/team292/vl6/VISIUM/femalereproductiveaxis_mese_cytassist_downsampled_fitted_values_tradeseq.csv')


hoxa5 <- plotSmoothers(sce, counts, gene = 'HOXA5', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA5") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa5

hoxa9 <- plotSmoothers(sce, counts, gene = 'HOXA9', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA9") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa9

hoxa10 <- plotSmoothers(sce, counts, gene = 'HOXA10', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA5") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa10

hoxc6 <- plotSmoothers(sce, counts, gene = 'HOXC6', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA5") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxc6

hoxa11 <- plotSmoothers(sce, counts, gene = 'HOXA11', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA5") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa11

hoxa7 <- plotSmoothers(sce, counts, gene = 'HOXA7', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA7") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa7

hoxa13 <- plotSmoothers(sce, counts, gene = 'HOXA13', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("HOXA13") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
hoxa13

p <- grid.arrange(hoxa5, hoxa9, hoxc5, hoxa10, HOXC6,  hoxa11, hoxa7,     hoxa13, 
                  nrow = 4, ncol = 2)

ggsave('/home/jovyan/RepTract/visium_femalereproductiveaxis_hox_mese_tradeseq.pdf',
       plot = p,
       width = 8, height = 13, dpi = 300)

BMP4 <- plotSmoothers(sce, counts, gene = 'BMP4', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("BMP4") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
BMP4

IGF1 <- plotSmoothers(sce, counts, gene = 'IGF1', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("IGF1") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
IGF1

SOX17 <- plotSmoothers(sce, counts, gene = 'SOX17', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("SOX17") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light() + ylim(0, 3.5)
SOX17

CSF1R <- plotSmoothers(sce, counts, gene = 'CSF1R', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("CSF1R") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light() + ylim(0, 3.2)
CSF1R

PNOC <- plotSmoothers(sce, counts, gene = 'PNOC', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("PNOC") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light() + ylim(0, 3.7)
PNOC

p2 <- grid.arrange(PNOC, SOX17, CSF1R,
                  nrow = 3, ncol = 1)

ggsave('/home/jovyan/RepTract/visium_femalereproductiveaxis_spurious_mese_tradeseq.pdf',
       plot = p2,
       width = 4, height = 9, dpi = 300)

TNC <- plotSmoothers(sce, counts, gene = 'TNC', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("TNC") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
TNC

SRD5A2 <- plotSmoothers(sce, counts, gene = 'SRD5A2', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("SRD5A2") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
SRD5A2


GDF10 <- plotSmoothers(sce, counts, gene = 'GDF10', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("GDF10") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
GDF10

PTGES <- plotSmoothers(sce, counts, gene = 'PTGES', alpha = 0, 
                       xlab = "Female Reproductive Axis",
  ylab = "Log(expression + 1)", border = FALSE) + ggtitle("PTGES") + theme(title = element_text(size = 20, face = 'bold'), axis.title=element_text(size=20), axis.text=element_text(size=20), legend.text = element_text(size=20),
     legend.title = element_text(size=20, face = "bold"))+ theme_light()
PTGES
